+++ ocdskit/cli/__main__.py
@@ -15,9 +15,12 @@ COMMAND_MODULES = (
     'ocdskit.cli.commands.indent',
     'ocdskit.cli.commands.mapping_sheet',
     'ocdskit.cli.commands.measure',
+    'ocdskit.cli.commands.package_releases',
     'ocdskit.cli.commands.schema_report',
     'ocdskit.cli.commands.schema_strict',
     'ocdskit.cli.commands.set_closed_codelist_enums',
+    'ocdskit.cli.commands.split_record_packages',
+    'ocdskit.cli.commands.split_release_packages',
     'ocdskit.cli.commands.tabulate',
     'ocdskit.cli.commands.validate',
 )
@@ -26,6 +29,7 @@ COMMAND_MODULES = (
 def main():
     parser = argparse.ArgumentParser(description='Open Contracting Data Standard CLI')
     parser.add_argument('--encoding', help='the file encoding')
+    parser.add_argument('--ascii', help='print escape sequences instead of UTF-8 characters', action='store_true')
     parser.add_argument('--pretty', help='pretty print output', action='store_true')
 
     subparsers = parser.add_subparsers(dest='subcommand')