+++ ocdskit/cli/commands/combine_release_packages.py
@@ -1,4 +1,3 @@
-import json
 from collections import OrderedDict
 
 from .base import BaseCommand
@@ -12,26 +11,12 @@ class Command(BaseCommand):
         output = OrderedDict([('extensions', OrderedDict()), ('releases', [])])
 
         for line in self.buffer():
-            package = json.loads(line, object_pairs_hook=OrderedDict)
+            package = self.json_loads(line)
 
-            # Use sample metadata.
-            output['uri'] = package['uri']
-            output['publishedDate'] = package['publishedDate']
-            output['publisher'] = package['publisher']
-
-            if 'extensions' in package:
-                # Python has no OrderedSet, so we use OrderedDict to keep extensions in order without duplication.
-                output['extensions'].update(dict.fromkeys(package['extensions'], True))
-
-            for field in ('license', 'publicationPolicy', 'version'):
-                if field in package:
-                    output[field] = package[field]
+            self._update_package_metadata(output, package)
 
             output['releases'].extend(package['releases'])
 
-        if output['extensions']:
-            output['extensions'] = list(output['extensions'])
-        else:
-            del output['extensions']
+        self._set_extensions_metadata(output)
 
         self.print(output)