+++ tests/commands/test_split_record_packages.py
@@ -1,0 +1,16 @@
+import sys
+from io import BytesIO, StringIO, TextIOWrapper
+from unittest.mock import patch
+
+from ocdskit.cli.__main__ import main
+from tests import read
+
+
+def test_command(monkeypatch):
+    stdin = read('realdata/record-package_package.json', 'rb')
+
+    with patch('sys.stdin', TextIOWrapper(BytesIO(stdin))), patch('sys.stdout', new_callable=StringIO) as actual:
+        monkeypatch.setattr(sys, 'argv', ['ocdskit', 'split-record-packages', '1'])
+        main()
+
+    assert actual.getvalue() == read('realdata/record-package_split.json')