+++ CHANGELOG.md
@@ -1,16 +1,24 @@
 # Changelog
 
-# 0.0.3
+# 0.0.3 (2018-11-01)
 
 New options:
 
-* compile: `--versioned`
+* compile: `--package`, `--versioned`
+
+New commands:
+
+* package-releases
+* split-record-packages
+* split-release-packages
 
 Other changes:
 
 * Add helpful error messages if:
   * the input is not [line-delimited JSON](https://en.wikipedia.org/wiki/JSON_streaming) data;
   * the input to the `indent` command is not valid JSON.
+* Change default behavior to print UTF-8 characters instead of escape sequences.
+* Add `--ascii` option to print escape sequences instead of UTF-8 characters.
 * Rename base exception class from `ReportError` to `OCDSKitError`.
 
 # 0.0.2 (2018-03-14)